requirements.txt
python-dotenv==1.0.1
requests==2.32.3
feedgen==1.0.0
tqdm==4.67.1
beautifulsoup4==4.12.3
matplotlib==3.9.2
pandas==2.2.2
openpyxl==3.1.5
